[Feat] #172 - 곧 마감인 공고 카드 클릭 시 스크랩 편집 기능 구현

+++ Terning-iOS/Terning-iOS/Source/Domain/Model/Home/UpcomingCard.swift
@@ -13,7 +13,7 @@ struct UpcomingCardModel: Codable {
 }
 
 struct UpcomingCard: Codable {
-    let internshipAnnouncementId: Double
+    let internshipAnnouncementId: Int
     let companyImage: String
     let dDay: String
     let title: String
